+++ src/Modules/Calculators/CalculatorsModule.cs
@@ -8,6 +8,7 @@ public static class CalculatorsModule
         services.AddScoped<IIrrfCalculator, IrrfCalculator>();
         services.AddScoped<ITerminationTaxCalculator, TerminationTaxCalculator>();
         services.AddScoped<NetSalaryCalculator>();
+        services.AddScoped<CltPjComparisonCalculator>();
         services.AddScoped<CalculationEngine>();
         services.AddSingleton<ICalculatorFieldProfileProvider, CalculatorFieldProfileProvider>();
         services.AddScoped<ICalculatorApplicationService, CalculatorApplicationService>();